@@ -5,7 +5,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ^3.5.0
+  sdk: ^3.8.0
 
 dependencies:
   bloc: ^9.0.0
@@ -23,13 +23,13 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   flutter_svg: ^2.2.0
-  freezed: ^3.2.0
   freezed_annotation: ^3.0.0
   get_it: ^8.0.3
   go_router: ^16.0.0
   google_maps_flutter: ^2.12.2
   google_sign_in: ^6.2.1
   intl: ^0.20.2
+  json_annotation: ^4.9.0
   multiple_result: ^5.2.0
   path: ^1.9.0
   path_provider: ^2.1.5
@@ -44,7 +44,9 @@ dev_dependencies:
   commitlint_cli: ^0.8.1
   flutter_test:
     sdk: flutter
+  freezed: ^3.2.0
   husky: ^0.1.7
+  json_serializable: ^6.9.0
   mockito: ^5.4.4
   very_good_analysis: ^8.0.0
 
