@@ -11,301 +11,258 @@ part of 'sign_in_state.dart';
 
 // dart format off
 T _$identity<T>(T value) => value;
-
 /// @nodoc
 mixin _$SignInState {
-  StateStatus get status;
-  Failure? get failure;
 
-  /// Create a copy of SignInState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $SignInStateCopyWith<SignInState> get copyWith =>
-      _$SignInStateCopyWithImpl<SignInState>(this as SignInState, _$identity);
+ RequestStatus get status; Failure? get failure;
+/// Create a copy of SignInState
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$SignInStateCopyWith<SignInState> get copyWith => _$SignInStateCopyWithImpl<SignInState>(this as SignInState, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is SignInState &&
-            (identical(other.status, status) || other.status == status) &&
-            (identical(other.failure, failure) || other.failure == failure));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, status, failure);
 
-  @override
-  String toString() {
-    return 'SignInState(status: $status, failure: $failure)';
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is SignInState&&(identical(other.status, status) || other.status == status)&&(identical(other.failure, failure) || other.failure == failure));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,status,failure);
+
+@override
+String toString() {
+  return 'SignInState(status: $status, failure: $failure)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $SignInStateCopyWith<$Res> {
-  factory $SignInStateCopyWith(
-          SignInState value, $Res Function(SignInState) _then) =
-      _$SignInStateCopyWithImpl;
-  @useResult
-  $Res call({StateStatus status, Failure? failure});
-}
+abstract mixin class $SignInStateCopyWith<$Res>  {
+  factory $SignInStateCopyWith(SignInState value, $Res Function(SignInState) _then) = _$SignInStateCopyWithImpl;
+@useResult
+$Res call({
+ RequestStatus status, Failure? failure
+});
 
+
+
+
+}
 /// @nodoc
-class _$SignInStateCopyWithImpl<$Res> implements $SignInStateCopyWith<$Res> {
+class _$SignInStateCopyWithImpl<$Res>
+    implements $SignInStateCopyWith<$Res> {
   _$SignInStateCopyWithImpl(this._self, this._then);
 
   final SignInState _self;
   final $Res Function(SignInState) _then;
 
-  /// Create a copy of SignInState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? status = null,
-    Object? failure = freezed,
-  }) {
-    return _then(_self.copyWith(
-      status: null == status
-          ? _self.status
-          : status // ignore: cast_nullable_to_non_nullable
-              as StateStatus,
-      failure: freezed == failure
-          ? _self.failure
-          : failure // ignore: cast_nullable_to_non_nullable
-              as Failure?,
-    ));
-  }
+/// Create a copy of SignInState
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? status = null,Object? failure = freezed,}) {
+  return _then(_self.copyWith(
+status: null == status ? _self.status : status // ignore: cast_nullable_to_non_nullable
+as RequestStatus,failure: freezed == failure ? _self.failure : failure // ignore: cast_nullable_to_non_nullable
+as Failure?,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [SignInState].
 extension SignInStatePatterns on SignInState {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_SignInState value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _SignInState() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _SignInState value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _SignInState() when $default != null:
+return $default(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_SignInState value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SignInState():
-        return $default(_that);
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _SignInState value)  $default,){
+final _that = this;
+switch (_that) {
+case _SignInState():
+return $default(_that);}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _SignInState value)?  $default,){
+final _that = this;
+switch (_that) {
+case _SignInState() when $default != null:
+return $default(_that);case _:
+  return null;
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_SignInState value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SignInState() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( RequestStatus status,  Failure? failure)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _SignInState() when $default != null:
+return $default(_that.status,_that.failure);case _:
+  return orElse();
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(StateStatus status, Failure? failure)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _SignInState() when $default != null:
-        return $default(_that.status, _that.failure);
-      case _:
-        return orElse();
-    }
-  }
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( RequestStatus status,  Failure? failure)  $default,) {final _that = this;
+switch (_that) {
+case _SignInState():
+return $default(_that.status,_that.failure);}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(StateStatus status, Failure? failure) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SignInState():
-        return $default(_that.status, _that.failure);
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( RequestStatus status,  Failure? failure)?  $default,) {final _that = this;
+switch (_that) {
+case _SignInState() when $default != null:
+return $default(_that.status,_that.failure);case _:
+  return null;
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(StateStatus status, Failure? failure)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SignInState() when $default != null:
-        return $default(_that.status, _that.failure);
-      case _:
-        return null;
-    }
-  }
 }
 
 /// @nodoc
+
 
 class _SignInState implements SignInState {
-  const _SignInState({this.status = StateStatus.initial, this.failure});
+  const _SignInState({this.status = RequestStatus.initial, this.failure});
+  
 
-  @override
-  @JsonKey()
-  final StateStatus status;
-  @override
-  final Failure? failure;
+@override@JsonKey() final  RequestStatus status;
+@override final  Failure? failure;
 
-  /// Create a copy of SignInState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$SignInStateCopyWith<_SignInState> get copyWith =>
-      __$SignInStateCopyWithImpl<_SignInState>(this, _$identity);
+/// Create a copy of SignInState
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$SignInStateCopyWith<_SignInState> get copyWith => __$SignInStateCopyWithImpl<_SignInState>(this, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _SignInState &&
-            (identical(other.status, status) || other.status == status) &&
-            (identical(other.failure, failure) || other.failure == failure));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, status, failure);
 
-  @override
-  String toString() {
-    return 'SignInState(status: $status, failure: $failure)';
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _SignInState&&(identical(other.status, status) || other.status == status)&&(identical(other.failure, failure) || other.failure == failure));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,status,failure);
+
+@override
+String toString() {
+  return 'SignInState(status: $status, failure: $failure)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class _$SignInStateCopyWith<$Res>
-    implements $SignInStateCopyWith<$Res> {
-  factory _$SignInStateCopyWith(
-          _SignInState value, $Res Function(_SignInState) _then) =
-      __$SignInStateCopyWithImpl;
-  @override
-  @useResult
-  $Res call({StateStatus status, Failure? failure});
-}
+abstract mixin class _$SignInStateCopyWith<$Res> implements $SignInStateCopyWith<$Res> {
+  factory _$SignInStateCopyWith(_SignInState value, $Res Function(_SignInState) _then) = __$SignInStateCopyWithImpl;
+@override @useResult
+$Res call({
+ RequestStatus status, Failure? failure
+});
 
+
+
+
+}
 /// @nodoc
-class __$SignInStateCopyWithImpl<$Res> implements _$SignInStateCopyWith<$Res> {
+class __$SignInStateCopyWithImpl<$Res>
+    implements _$SignInStateCopyWith<$Res> {
   __$SignInStateCopyWithImpl(this._self, this._then);
 
   final _SignInState _self;
   final $Res Function(_SignInState) _then;
 
-  /// Create a copy of SignInState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? status = null,
-    Object? failure = freezed,
-  }) {
-    return _then(_SignInState(
-      status: null == status
-          ? _self.status
-          : status // ignore: cast_nullable_to_non_nullable
-              as StateStatus,
-      failure: freezed == failure
-          ? _self.failure
-          : failure // ignore: cast_nullable_to_non_nullable
-              as Failure?,
-    ));
-  }
+/// Create a copy of SignInState
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? status = null,Object? failure = freezed,}) {
+  return _then(_SignInState(
+status: null == status ? _self.status : status // ignore: cast_nullable_to_non_nullable
+as RequestStatus,failure: freezed == failure ? _self.failure : failure // ignore: cast_nullable_to_non_nullable
+as Failure?,
+  ));
+}
+
+
 }
 
 // dart format on
