@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:multitec_app/core/di/service_locator.dart';
 import 'package:multitec_app/core/l10n/l10n.dart';
 import 'package:multitec_app/core/router/app_router.dart';
-import 'package:multitec_app/core/ui/theme/theme.dart';
+import 'package:multitec_app/core/ui/theme/app_theme.dart';
 import 'package:multitec_app/features/settings/presentation/cubits/locale_cubit.dart';
 import 'package:multitec_app/features/settings/presentation/cubits/theme_cubit.dart';
 import 'package:multitec_app/features/user/presentation/cubits/user_cubit.dart';
