@@ -1,4 +1,4 @@
-import 'package:multitec_app/core/ui/cubit/state_status.dart';
+import 'package:multitec_app/core/ui/cubit/request_status.dart';
 import 'package:multitec_app/core/utils/safe_cubit.dart';
 import 'package:multitec_app/features/auth/domain/usecases/sign_in_with_google_usecase.dart';
 import 'package:multitec_app/features/auth/presentation/cubit/sign_in_state.dart';
@@ -11,23 +11,15 @@ class SignInCubit extends SafeCubit<SignInState> {
   Future<void> signInWithGoogle() async {
     if (state.status.isLoading) return;
 
-    emit(state.copyWith(status: StateStatus.loading));
+    emit(state.copyWith(status: RequestStatus.loading));
 
     final result = await _signInWithGoogle.call();
 
     result.when(
-      (user) => emit(
-        state.copyWith(
-          status: StateStatus.loaded,
-          failure: null,
-        ),
-      ),
-      (failure) => emit(
-        state.copyWith(
-          status: StateStatus.error,
-          failure: failure,
-        ),
-      ),
+      (user) =>
+          emit(state.copyWith(status: RequestStatus.success, failure: null)),
+      (failure) =>
+          emit(state.copyWith(status: RequestStatus.failure, failure: failure)),
     );
   }
 }
