@@ -1,5 +1,6 @@
-import 'package:multitec_app/features/schedule/domain/models/paginated_result.dart';
+import 'package:multitec_app/features/schedule/domain/entities/paginated_result.dart';
 
+//TODO: Convertir a freezed?
 class PaginatedResultDto<T> {
   const PaginatedResultDto({
     required this.items,
