@@ -1,16 +1,13 @@
 import 'package:multiple_result/multiple_result.dart';
 import 'package:multitec_app/core/exceptions/failure.dart';
-import 'package:multitec_app/features/schedule/domain/models/paginated_result.dart';
-import 'package:multitec_app/features/schedule/domain/models/pagination_params.dart';
-import 'package:multitec_app/features/schedule/domain/models/schedule_item.dart';
+import 'package:multitec_app/features/schedule/domain/entities/paginated_result.dart';
+import 'package:multitec_app/features/schedule/domain/entities/pagination_params.dart';
+import 'package:multitec_app/features/schedule/domain/entities/schedule_item.dart';
 import 'package:multitec_app/features/schedule/domain/repositories/schedule_repository.dart';
 import 'package:multitec_app/features/user/domain/repositories/user_repository.dart';
 
 class GetJoinedScheduleItemsUseCase {
-  GetJoinedScheduleItemsUseCase(
-    this._scheduleRepository,
-    this._userRepository,
-  );
+  GetJoinedScheduleItemsUseCase(this._scheduleRepository, this._userRepository);
 
   final ScheduleRepository _scheduleRepository;
   final UserRepository _userRepository;
@@ -21,16 +18,16 @@ class GetJoinedScheduleItemsUseCase {
   }) async {
     final userResult = _userRepository.getUser();
 
-    return userResult.when(
-      (user) {
-        if (user == null) return const Result.error(GenericFailure());
+    return userResult.when((user) {
+      if (user == null) return const Result.error(GenericFailure());
 
-        return _scheduleRepository.getJoinedScheduleItems(
-          user.id,
-          PaginationParams(cursor: cursor, limit: limit),
-        );
-      },
-      Result.error,
-    );
+      return _scheduleRepository.getJoinedScheduleItems(
+        user.id,
+        PaginationParams(
+          cursor: cursor,
+          limit: limit ?? 20,
+        ), //TODO: Ver esto para no pasar 20 a pelo
+      );
+    }, Result.error);
   }
 }
