@@ -1,7 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:multitec_app/core/exceptions/failure.dart';
-import 'package:multitec_app/core/ui/cubit/state_status.dart';
-import 'package:multitec_app/features/schedule/domain/models/schedule_item.dart';
+import 'package:multitec_app/core/ui/cubit/request_status.dart';
+import 'package:multitec_app/features/schedule/domain/entities/schedule_item.dart';
 
 part 'joined_schedules_state.freezed.dart';
 
@@ -9,7 +9,7 @@ part 'joined_schedules_state.freezed.dart';
 sealed class JoinedSchedulesState with _$JoinedSchedulesState {
   const factory JoinedSchedulesState({
     @Default(<ScheduleItem>[]) List<ScheduleItem> items,
-    @Default(StateStatus.initial) StateStatus status,
+    @Default(RequestStatus.initial) RequestStatus status,
     @Default(true) bool hasMore,
     Failure? failure,
   }) = _JoinedSchedulesState;
