@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:multitec_app/core/ui/cubit/state_status.dart';
+import 'package:multitec_app/core/ui/cubit/request_status.dart';
 import 'package:multitec_app/features/auth/domain/enums/autentication_status.dart';
 import 'package:multitec_app/features/auth/domain/services/auth_service.dart';
 import 'package:multitec_app/features/user/domain/repositories/user_repository.dart';
@@ -8,7 +8,7 @@ import 'package:multitec_app/features/user/presentation/cubits/user_state.dart';
 
 class UserCubit extends Cubit<UserState> {
   UserCubit(this._userRepository, this._authService)
-      : super(const UserState()) {
+    : super(const UserState()) {
     _onAuthStatusChanged = () {
       final status = _authService.status;
       if (status == AuthenticationStatus.authenticated) {
@@ -28,24 +28,20 @@ class UserCubit extends Cubit<UserState> {
   Future<void> getUser() async {
     if (state.status.isLoading) return;
 
-    emit(state.copyWith(status: StateStatus.loading));
+    emit(state.copyWith(status: RequestStatus.loading));
 
     final result = _userRepository.getUser();
 
     result.when(
       (user) => emit(
         state.copyWith(
-          status: StateStatus.loaded,
+          status: RequestStatus.success,
           user: user,
           failure: null,
         ),
       ),
-      (failure) => emit(
-        state.copyWith(
-          status: StateStatus.error,
-          failure: failure,
-        ),
-      ),
+      (failure) =>
+          emit(state.copyWith(status: RequestStatus.failure, failure: failure)),
     );
   }
 
