@@ -1,14 +1,14 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:multitec_app/core/exceptions/failure.dart';
-import 'package:multitec_app/core/ui/cubit/state_status.dart';
-import 'package:multitec_app/features/user/domain/models/user.dart';
+import 'package:multitec_app/core/ui/cubit/request_status.dart';
+import 'package:multitec_app/features/user/domain/entities/user.dart';
 
 part 'user_state.freezed.dart';
 
 @freezed
 sealed class UserState with _$UserState {
   const factory UserState({
-    @Default(StateStatus.initial) StateStatus status,
+    @Default(RequestStatus.initial) RequestStatus status,
     User? user,
     Failure? failure,
   }) = _UserState;
